Station Captain- Separated out input validators from user interaction
